refactor(player-web-demo): split main.ts into feature modules
Extract the inline HTML template, type declarations, and standalone utility
functions out of the 3,979-line entry point so the bootstrap is no longer
dominated by markup and helpers. PlayerWebDemoApp itself is unchanged.

- dom-template.ts: app.innerHTML markup (DEMO_APP_HTML)
- types.ts:        PlayerWebDemoElements / DemoGuiState / SkinFamilyOverride
                   / SkinFamilySceneKind / ChartImageTexture
- compat-panel.ts: renderBrowserCompatPanel + buildCompatRow
- readtext-overlay.ts: decodeText + showReadtextOverlay
- demo-utils.ts:   phaseLabels, sanitizeFilenameStem, playSkinTypeForVariant

main.ts drops from 3,979 to 3,201 lines (-778). No behavior change; the
moved declarations are imported back verbatim.
